PR changes

+++ migrations/app/schema/20210414170143_add_FSC_service_item_param_keys.up.sql
@@ -2,7 +2,7 @@ INSERT INTO service_item_param_keys
 (id, key,description,type,origin,created_at,updated_at)
 VALUES
 ('117da2f5-fff0-41e0-bba1-837124373098', 'FSCMultiplier', 'Cost multiplier multiplied by FSCPriceDifferenceInCents based on the FSCWeightBasedDistanceMultiplier and Distance', 'DECIMAL', 'PRICER', now(), now()),
-('6ba0aeca-19f8-4247-a317-fffa81c5d5c1', 'FSCPriceDifferenceInCents', 'Difference in price between the weekly national average EIA fuel price and the base GHC diesel fuel price of $2.50', 'DECIMAL', 'PRICER', now(), now());
+('6ba0aeca-19f8-4247-a317-fffa81c5d5c1', 'FSCPriceDifferenceInCents', 'Difference in price between the weekly national average EIA fuel price and the base GHC diesel fuel price of $2.50', 'INTEGER', 'PRICER', now(), now());
 
 INSERT INTO service_params
 (id,service_id,service_item_param_key_id,created_at,updated_at)
